refactor(pipeline): decompose the listener pipeline god-file by role (WS-D)
Split internal/alerts/pipeline/listener_pipeline.go (654 LOC) into
role-based files in the same package — pure move: pipeline lifecycle stays
in listener_pipeline.go; rule loading/snapshotting moves to
listener_rules.go; event evaluation moves to listener_eval.go.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
